Fixed reading and writing at unaligned offsets on BSD
FreeBSD (and other BSDs) do not support reading/writing at
random offsets within block devices. The underlying read(2) call
will fail with EINVAL if an attempt is made to read/write at
unaligned boundaries.

This patch modifies contains_data() to read at the pagesize
boundary, and then check at the corresponding offset for the
search parameters. It reads a mulitple of pagesize as well. It
effectively reads (roundUp(offset + bytes) - roundDown(offset))
bytes starting from roundDown(offset).

For writes in write_data(), first the existing content from
roundDown(offset) to roundUp(offset + bytes) is read into a
buffer, and then the contents of that buffer are patched with the
data we wish to write to disk, and then the entire buffer is
written back. All reads and writes are of pagesize multiple.